+++ integration_test/ownership/manager_test.exs
@@ -49,6 +49,7 @@ defmodule ManagerTest do
 
   test "connection may be shared with other processes" do
     {:ok, pool, _opts} = start_pool()
+    {:ok, pool_2, _opts} = start_pool()
     parent = self()
 
     Task.await(
@@ -63,6 +64,12 @@ defmodule ManagerTest do
     assert Ownership.ownership_allow(pool, self(), self(), []) ==
              {:already, :owner}
 
+    :ok = Ownership.ownership_checkout(pool_2, [])
+
+    assert Ownership.ownership_allow(pool_2, self(), self(), unallow_existing: true) == :ok
+
+    assert Ownership.ownership_allow(pool_2, self(), self(), []) == {:already, :allowed}
+
     Task.await(
       async_no_callers(fn ->
         assert Ownership.ownership_allow(pool, parent, self(), []) ==
@@ -72,6 +79,13 @@ defmodule ManagerTest do
                  {:already, :allowed}
 
         assert Ownership.ownership_checkin(pool, []) == :not_owner
+
+        # Test that we can switch our 'allow' to another pool and back again
+        assert Ownership.ownership_allow(pool_2, parent, self(), unallow_existing: true) == :ok
+
+        assert Ownership.ownership_allow(pool_2, parent, self(), []) == {:already, :allowed}
+
+        assert Ownership.ownership_allow(pool, parent, self(), unallow_existing: true) == :ok
 
         parent = self()
 
@@ -309,7 +323,6 @@ defmodule ManagerTest do
     Task.await(task)
   end
 
-  @tag :requires_callers
   test "performs callers checkout on manual mode" do
     {:ok, pool, opts} = start_pool()
     assert Ownership.ownership_mode(pool, :manual, []) == :ok
@@ -324,7 +337,22 @@ defmodule ManagerTest do
     |> Task.await()
   end
 
-  @tag :requires_callers
+  test "performs callers checkout on manual mode through allowance" do
+    {:ok, pool, opts} = start_pool()
+    assert Ownership.ownership_mode(pool, :manual, []) == :ok
+
+    :ok = Ownership.ownership_checkout(pool, [])
+    assert_checked_out(pool, opts)
+    parent = self()
+
+    async_no_callers(fn ->
+      assert Ownership.ownership_allow(pool, parent, self(), []) == :ok
+      assert_checked_out(pool, opts)
+      Task.async(fn -> assert_checked_out(pool, opts) end) |> Task.await()
+    end)
+    |> Task.await()
+  end
+
   test "does not perform callers checkout on auto mode" do
     {:ok, agent} =
       A.start_link([{:ok, :state}, {:ok, :state}] ++ List.duplicate({:idle, :state}, 10))
@@ -608,6 +636,7 @@ defmodule ManagerTest do
       {:ok, :state},
       {:idle, :state},
       :oops,
+      :ok,
       fn opts ->
         send(opts[:parent], :reconnected)
         {:ok, :state}