+++ lib/db_connection/ownership/manager.ex
@@ -3,6 +3,7 @@ defmodule DBConnection.Ownership.Manager do
   use GenServer
   require Logger
   alias DBConnection.Ownership.Proxy
+  alias DBConnection.Util
 
   @timeout 5_000
 
@@ -56,7 +57,14 @@ defmodule DBConnection.Ownership.Manager do
           :ok | {:already, :owner | :allowed} | :not_found
   def allow(manager, parent, allow, opts) do
     timeout = Keyword.get(opts, :timeout, @timeout)
-    GenServer.call(manager, {:allow, parent, allow}, timeout)
+    passed_opts = Keyword.take(opts, [:unallow_existing])
+    GenServer.call(manager, {:allow, parent, allow, passed_opts}, timeout)
+  end
+
+  @spec get_connection_metrics(GenServer.server()) ::
+          {:ok, [DBConnection.Pool.connection_metrics()]} | :error
+  def get_connection_metrics(manager) do
+    GenServer.call(manager, :get_connection_metrics, :infinity)
   end
 
   ## Callbacks
@@ -68,7 +76,13 @@ defmodule DBConnection.Ownership.Manager do
     ets =
       case Keyword.fetch(owner_opts, :name) do
         {:ok, name} when is_atom(name) ->
-          :ets.new(name, [:set, :named_table, :protected, read_concurrency: true])
+          :ets.new(name, [
+            :set,
+            :named_table,
+            :protected,
+            read_concurrency: true,
+            decentralized_counters: true
+          ])
 
         _ ->
           nil
@@ -98,6 +112,31 @@ defmodule DBConnection.Ownership.Manager do
   end
 
   @impl true
+  def handle_call(:get_connection_metrics, _from, %{pool: pool, owners: owners, log: log} = state) do
+    pool_metrics = DBConnection.ConnectionPool.get_connection_metrics(pool)
+
+    proxy_metrics =
+      owners
+      |> Enum.map(fn {_, {proxy, _, _}} ->
+        try do
+          GenServer.call(proxy, :get_connection_metrics)
+        catch
+          :exit, reason ->
+            if log do
+              Logger.log(
+                log,
+                "Caught :exit while calling :get_connection_metrics due to #{inspect(reason)}"
+              )
+            end
+
+            nil
+        end
+      end)
+      |> Enum.reject(&is_nil/1)
+
+    {:reply, pool_metrics ++ proxy_metrics, state}
+  end
+
   def handle_call(:pool, _from, %{pool: pool} = state) do
     {:reply, pool, state}
   end
@@ -133,15 +172,24 @@ defmodule DBConnection.Ownership.Manager do
     {:reply, reply, state}
   end
 
-  def handle_call({:allow, caller, allow}, _from, %{checkouts: checkouts} = state) do
-    if kind = already_checked_out(checkouts, allow) do
+  def handle_call({:allow, caller, allow, opts}, _from, %{checkouts: checkouts} = state) do
+    unallow_existing = Keyword.get(opts, :unallow_existing, false)
+    kind = already_checked_out(checkouts, allow)
+
+    if !unallow_existing && kind do
       {:reply, {:already, kind}, state}
     else
       case Map.get(checkouts, caller, :not_found) do
         {:owner, ref, proxy} ->
+          state =
+            if unallow_existing, do: owner_unallow(state, caller, allow, ref, proxy), else: state
+
           {:reply, :ok, owner_allow(state, caller, allow, ref, proxy)}
 
         {:allowed, ref, proxy} ->
+          state =
+            if unallow_existing, do: owner_unallow(state, caller, allow, ref, proxy), else: state
+
           {:reply, :ok, owner_allow(state, caller, allow, ref, proxy)}
 
         :not_found ->
@@ -273,6 +321,24 @@ defmodule DBConnection.Ownership.Manager do
     state
   end
 
+  defp owner_unallow(%{ets: ets, log: log} = state, caller, unallow, ref, proxy) do
+    if log do
+      Logger.log(log, fn ->
+        [inspect(unallow), " was unallowed by ", inspect(caller), " on proxy ", inspect(proxy)]
+      end)
+    end
+
+    state = update_in(state.checkouts, &Map.delete(&1, unallow))
+
+    state =
+      update_in(state.owners[ref], fn {proxy, caller, allowed} ->
+        {proxy, caller, List.delete(allowed, unallow)}
+      end)
+
+    ets && :ets.delete(ets, {unallow, proxy})
+    state
+  end
+
   defp owner_down(%{ets: ets, log: log} = state, ref) do
     case get_and_update_in(state.owners, &Map.pop(&1, ref)) do
       {{proxy, caller, allowed}, state} ->
@@ -329,7 +395,7 @@ defmodule DBConnection.Ownership.Manager do
 
   defp not_found({pid, _} = from) do
     msg = """
-    cannot find ownership process for #{inspect(pid)}.
+    cannot find ownership process for #{Util.inspect_pid(pid)}.
 
     When using ownership, you must manage connections in one
     of the four ways: