+++ lib/db_connection/ownership/proxy.ex
@@ -2,6 +2,7 @@ defmodule DBConnection.Ownership.Proxy do
   @moduledoc false
 
   alias DBConnection.Holder
+  alias DBConnection.Util
   use GenServer, restart: :temporary
 
   @time_unit 1000
@@ -21,6 +22,8 @@ defmodule DBConnection.Ownership.Proxy do
 
   @impl true
   def init({caller, pool, pool_opts}) do
+    Util.set_label("db_ownership_proxy")
+
     pool_opts =
       pool_opts
       |> Keyword.put(:timeout, :infinity)
@@ -58,13 +61,13 @@ defmodule DBConnection.Ownership.Proxy do
 
   @impl true
   def handle_info({:DOWN, ref, _, pid, _reason}, %{owner: {_, ref}} = state) do
-    down("owner #{inspect(pid)} exited", state)
+    down("owner #{Util.inspect_pid(pid)} exited", state)
   end
 
   def handle_info({:timeout, deadline, {_ref, holder, pid, len}}, %{holder: holder} = state) do
     if Holder.handle_deadline(holder, deadline) do
       message =
-        "client #{inspect(pid)} timed out because " <>
+        "client #{Util.inspect_pid(pid)} timed out because " <>
           "it queued and checked out the connection for longer than #{len}ms"
 
       down(message, state)
@@ -78,7 +81,7 @@ defmodule DBConnection.Ownership.Proxy do
         %{ownership_timer: timer} = state
       ) do
     message =
-      "owner #{inspect(pid)} timed out because " <>
+      "owner #{Util.inspect_pid(pid)} timed out because " <>
         "it owned the connection for longer than #{timeout}ms (set via the :ownership_timeout option)"
 
     # We don't invoke down because this is always a disconnect, even if there is no client.
@@ -150,7 +153,7 @@ defmodule DBConnection.Ownership.Proxy do
   end
 
   def handle_info({:"ETS-TRANSFER", holder, pid, ref}, %{holder: holder, owner: {_, ref}} = state) do
-    down("client #{inspect(pid)} exited", state)
+    down("client #{Util.inspect_pid(pid)} exited", state)
   end
 
   @impl true
@@ -169,6 +172,26 @@ defmodule DBConnection.Ownership.Proxy do
       end
 
     down(message, state)
+  end
+
+  @impl true
+  def handle_call(
+        :get_connection_metrics,
+        _,
+        %{queue: queue, holder: holder, client: client} = state
+      ) do
+    connection_metrics = %{
+      source: {:proxy, self()},
+      ready_conn_count:
+        if is_nil(holder) or not is_nil(client) do
+          0
+        else
+          1
+        end,
+      checkout_queue_length: :queue.len(queue)
+    }
+
+    {:reply, connection_metrics, state}
   end
 
   defp checkout({pid, ref} = from, %{holder: holder} = state) do